:zap: Build and push tagged versions to RubyGems
Automatically pushing gem versions on merge to master makes it difficult
to merge changes - they explicitly will require the gem version to be
bumped, which is not always required. Making the process of building and
pushing gems to RubyGems reliant on a tag in version control also means
that any version of this gem will be explicitly linked to a version
control tag.